Tnn torch subgraph gpu mem (#1644)
* update trt network gpu mem share

* add trt network share memory set

* update shared memory manager

* use torch memory for trt network && support upsample in rcnn

* support bilinear upsample

* fix shared memory manager alloc

* reduce scratch memory to 512m

* fix torch graph md5 randomness

* Revert "reduce scratch memory to 512m"

This reverts commit 81397d465fbaacd23b841406242951f0f21fe2f2.

* add group norm support for torch

+++ include/tnn/core/instance.h
@@ -55,7 +55,7 @@ public:
     Status DeInit();
 
     //  return memory bytes required for forward
-    Status GetForwardMemorySize(int& memory_size);
+    Status GetForwardMemorySize(size_t& memory_size);
 
     //  set memory to tnn instance. if success, return status code zero.
     //  only instance created with SHARE_MEMORY_MODE_SET_FROM_EXTERNAL can be set from external.
